[refactor]: hooks 코드 리팩토링
- hooks로import할 수 있게 함
- hooks의 userInfo -> user로 변경
