+++ src/hooks/mutations/index.ts
@@ -3,3 +3,15 @@ export { usePostAnswerMutation } from './question'
 
 // answer
 export { useDeleteAnswerMutation, useUpdateAnswerMutation } from './answer'
+
+// user
+
+export {
+  usePatchUserInfoMutation,
+  useDeleteUser,
+  useLogoutMutation,
+  useCheckNicknameMutation,
+  useUnsubscribeNewsLetterMutation,
+  usePostUserNotiQuery,
+  useUpdateUserPushToken
+} from './user'
