--- src/hooks/queries/checkId.ts
@@ -1,16 +1,0 @@
-import { apis } from 'apis/apis'
-
-export const useCheckIdQuery = (loginType: string, socialId: string, idToken?: string) => {
-  return {
-    queryKey: ['userToken'],
-    queryFn: async () => await apis.checkUserInfo(loginType, socialId, idToken),
-    onSuccess: (res: any) => {
-      return res
-    },
-    onError: (err: any) => {
-      return err
-    },
-    enabled: socialId.length > 0 || (idToken !== undefined && idToken?.length > 0),
-    retry: 0
-  }
-}
