+++ src/hooks/queries/index.ts
@@ -1,7 +1,16 @@
-export { useTestQuery } from './test'
-export { useKakaoQueries, useAppleQueries } from './oauth'
-export { useCheckIdQuery } from './checkId'
-export { useSignUpQuery } from './signup'
+// question
+export { useGetAnswerQuery, useGetQuestionQuery, useGetAnswerRecordCountQuery } from './question'
+
+// answer
+export { useGetMyAnswerListQuery } from './answer'
+
+// newsletter
 export { useNewsLetterQuery } from './newsletter'
-export { useGetAnswerQuery } from './question'
-export { useGetQuestionQuery } from './question'
+
+// oauth
+export { useKakaoQueries, useAppleQueries, useCheckIdQuery, useSignUpQuery } from './oauth'
+
+// user
+export { useGetUserInfoQuery, useGetUserNotiQuery, useGetUserPushToken } from './user'
+
+export { useTestQuery } from './test'
