+++ src/hooks/queries/oauth.ts
@@ -1,13 +1,25 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { useState } from 'react'
-import { useQueries } from 'react-query'
+import { useQueries, useQuery } from 'react-query'
 
 // apis
 import { apis } from 'apis/apis'
-import type { IKakaoTokenData } from 'apis/userApi'
+import type { IJoinResponseDto, IKakaoTokenData } from 'apis/userApi'
 
-// hookss
-import { useCheckIdQuery } from './checkId'
+export const useCheckIdQuery = (loginType: string, socialId: string, idToken?: string) => {
+  return {
+    queryKey: ['userToken'],
+    queryFn: async () => await apis.checkUserInfo(loginType, socialId, idToken),
+    onSuccess: (res: any) => {
+      return res
+    },
+    onError: (err: any) => {
+      return err
+    },
+    enabled: socialId.length > 0 || (idToken !== undefined && idToken?.length > 0),
+    retry: 0
+  }
+}
 
 export const useKakaoQueries = ({ client_id, redirect_uri, code }: IKakaoTokenData) => {
   const [accessToken, setAccessToken] = useState('')
@@ -42,4 +54,40 @@ export const useKakaoQueries = ({ client_id, redirect_uri, code }: IKakaoTokenDa
 
 export const useAppleQueries = ({ idToken }: { idToken: string }) => {
   return useQueries([useCheckIdQuery('apple', '', idToken)])
+}
+
+export const useSignUpQuery = ({
+  allowNotification,
+  birthDate,
+  gender,
+  job,
+  jobInfo,
+  mbti,
+  nickname,
+  purpose,
+  socialId,
+  socialLoginType,
+  pushToken
+}: IJoinResponseDto) => {
+  return useQuery(
+    ['signup'],
+    async () =>
+      await apis.signup({
+        allowNotification,
+        birthDate,
+        gender,
+        job,
+        jobInfo,
+        mbti,
+        nickname,
+        purpose,
+        socialId,
+        socialLoginType,
+        pushToken
+      }),
+    {
+      enabled: false,
+      retry: 0
+    }
+  )
 }
